test: send_slack_message
