+++ prowler/__main__.py
@@ -172,7 +172,7 @@ def prowler():
     stats = extract_findings_statistics(findings)
 
     if "SLACK_API_TOKEN" in os.environ and "SLACK_CHANNEL_ID" in os.environ:
-        send_slack_message(
+        _ = send_slack_message(
             os.environ["SLACK_API_TOKEN"],
             os.environ["SLACK_CHANNEL_ID"],
             stats,
